zstdgrep: fix behavior of -f option
-e and -f both eliminate the need for a pattern to be specified as the
first positional argument of grep, and should be treated the same way.

To make the code more generalizable, convert any pattern from the
positional arguments to an -e option rather than vice versa.

Note: this does not correctly handle multiple -f options (but this was
already an issue with -e).